# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Portal API Documentation",
        "description": "API for managing Job Offers",
        "contact": {
            "email": "support@musicbox.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Job Portal API Server"
        }
    ],
    "paths": {
...
        "/api/candidatures": {
            "post": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Soumettre une nouvelle candidature",
                "description": "Enregistre une nouvelle candidature pour une offre d'emploi spécifique.",
                "operationId": "62704344c7a43c99304e8d0a1c5e956e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "job_id",
                                    "full_name",
                                    "can_email",
                                    "phone_number"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID de l'utilisateur qui postule.",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "job_id": {
                                        "description": "ID de l'offre d'emploi concernée.",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "full_name": {
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "can_email": {
                                        "description": "Email du candidat (champ `can_email` dans la requête).",
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@candidat.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0612345678"
                                    },
                                    "motivation": {
                                        "type": "string",
                                        "example": "Je suis très intéressé par ce poste.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidature créée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "job_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "full_name": {
                                            "type": "string",
                                            "example": "Jean Dupont"
                                        },
                                        "can_email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "jean.dupont@candidat.com"
                                        },
                                        "phone_number": {
                                            "type": "string",
                                            "example": "0612345678"
                                        },
                                        "motivation": {
                                            "type": "string",
                                            "example": "Je suis très intéressé par ce poste.",
                                            "nullable": true
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation des données."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/candidatures/user/{user_id}": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Liste des candidatures par utilisateur",
                "description": "Récupère toutes les candidatures soumises par un utilisateur spécifique (ID). L'accès doit être restreint à l'utilisateur lui-même ou à un admin/employeur (Logique d'autorisation supposée).",
                "operationId": "77c12fcbf9df5ec1bd23701103ba12d9",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID de l'utilisateur dont on veut récupérer les candidatures.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des candidatures récupérées avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "job_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "full_name": {
                                                "type": "string",
                                                "example": "Jean Dupont"
                                            },
                                            "can_email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "jean.dupont@candidat.com"
                                            },
                                            "phone_number": {
                                                "type": "string",
                                                "example": "0612345678"
                                            },
                                            "motivation": {
                                                "type": "string",
                                                "example": "Motivation pour le poste.",
                                                "nullable": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune candidature trouvée pour cet utilisateur.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Aucune candidature trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/candidatures/job/{job_id}": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Liste des candidatures par offre d'emploi",
                "description": "Récupère toutes les candidatures soumises pour une offre d'emploi spécifique (ID). L'accès est limité aux rôles 'admin' et 'emp'.",
                "operationId": "a3f96e8c40e833db058f906feb80117c",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "ID de l'offre d'emploi pour laquelle on veut récupérer les candidatures.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des candidatures récupérées avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "job_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "full_name": {
                                                "type": "string",
                                                "example": "Jean Dupont"
                                            },
                                            "can_email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "jean.dupont@candidat.com"
                                            },
                                            "phone_number": {
                                                "type": "string",
                                                "example": "0612345678"
                                            },
                                            "motivation": {
                                                "type": "string",
                                                "example": "Motivation pour le poste.",
                                                "nullable": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Interdit (Non autorisé).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "not exiscte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune candidature trouvée pour cette offre d'emploi.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Aucune candidature trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Bearer Token Authentication",
                "in": "header",
                "bearerFormat": "Sanctum Token",
                "scheme": "Bearer"
            }
        }
    },
...
}